last commit todos by user

+++ src/app/api/todos/[id]/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 import * as yup from "yup";
@@ -20,11 +21,20 @@ const putSchema = yup.object({
   complete: yup.boolean().optional(), // ! Mostrar algo interesante
 });
 export async function PUT(request: Request, { params }: Args) {
+  const user = await getUserSessionServer();
+  if (!user || !user.id)
+    return NextResponse.json({ error: "no authenticated" }, { status: 401 });
   const { id } = params;
   const todo = await prisma.todo.findFirst({ where: { id } });
 
   if (!todo)
     return NextResponse.json({ message: "Not Found" }, { status: 404 });
+
+  if (todo.userId !== user.id)
+    return NextResponse.json(
+      { message: "don't have permissions" },
+      { status: 403 }
+    );
   try {
     const { description, complete } = await putSchema.validate(
       await request.json()
@@ -41,8 +51,11 @@ export async function PUT(request: Request, { params }: Args) {
 }
 
 export async function DELETE(request: Request, { params }: Args) {
+  const user = await getUserSessionServer();
+  if (!user || !user.id)
+    return NextResponse.json({ error: "no authenticated" }, { status: 401 });
   const { id } = params;
-  const todo = await prisma.todo.delete({ where: { id } });
+  const todo = await prisma.todo.delete({ where: { id, userId: user.id } });
   if (!todo)
     return NextResponse.json({ message: "Not Found" }, { status: 404 });
